requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
confluent-kafka==2.3.0
idna==3.7
pip==23.0.1
PyYAML==6.0.1
requests==2.31.0
setuptools==65.5.0
urllib3==2.2.1